[code] Minor changes, cleaning.

+++ code/app/sensors.h
@@ -35,7 +35,7 @@ void reEnableInputCapture(TIMCAPDriver *timcapp);
 
 uint8_t calculateTpFromMillivolt(uint16_t AnMin, uint16_t AnMax, uint16_t AnVal);
 uint8_t calculateAFRFromMillivolt(uint16_t afrMin, uint16_t afrMax, uint16_t AnVal);
-uint8_t calculateRpmFromHertz(uint16_t freq, uint16_t ratio);
+uint16_t calculateRpmFromHertz(uint16_t freq, uint16_t ratio);
 uint8_t calculateKnockIntensity(uint16_t tgtFreq, uint16_t ratio, uint16_t smplFreq, uint8_t* buffer, uint16_t size);
 
 #endif
